yeah

+++ Server/StorageNode.py
@@ -9,7 +9,7 @@ from Server.Server import Server
 from Server.baseClient import baseClient
 from MessageAssembler.ResponseAssembler import ResponseAssembler
 from MessageAssembler.RequestAssembler import RequestAssembler
-from Server.DirectoryServer import DirectoryServer
+from Exception.ServerError import ServerError
 
 
 class ThreadedTCPRequestHandler(socketserver.BaseRequestHandler):
@@ -22,20 +22,20 @@ class ThreadedTCPRequestHandler(socketserver.BaseRequestHandler):
         Process coming in TCP message and send response
         :return:
         '''
-        data = str(SocketMessageManager.recvMessage(self.request, self.server.getServer().statisticHelper), 'utf-8')
-        startTime = time.time() * 1000
-        if self.server.getServer().output == 'debug':
-            print("Server {} Received: {}".format(self.server.getServer().id, data))
-        response = self.processRequest(json.loads(data), self.server.getServer())
-        self.request.settimeout(0.5)
         try:
+            data = str(SocketMessageManager.recvMessage(self.request, self.server.getServer().statisticHelper), 'utf-8')
+            startTime = time.time() * 1000
+            if self.server.getServer().output == 'debug':
+                print("Server {} Received: {}".format(self.server.getServer().id, data))
+            response = self.processRequest(json.loads(data), self.server.getServer())
+            self.request.settimeout(0.5)
             SocketMessageManager.sendMessage(self.request, bytes(response, 'utf-8'),
                                              self.server.getServer().statisticHelper)
             self.server.getServer().statisticHelper.computeAverageResponseTime(startTime)
-        except socket.timeout:
+        except (socket.timeout, ConnectionRefusedError):
             print("Server {} timeout".format(self.request.address))
             print("Request Fault Tolerance Schema")
-            raise socket.timeout
+            raise ServerError('Server Error')
         if self.server.getServer().output == 'debug':
             print("Server {} send: {}".format(self.server.getServer().id, response))
 
@@ -56,7 +56,9 @@ class ThreadedTCPRequestHandler(socketserver.BaseRequestHandler):
             fileName = request['fileName']
             return server.createReadFileResponse(fileName)
         elif requestHead == 'getFileListFromNodeRequest':
-            return server.assembleGetFileListFromNodeResponse()
+            return server.createGetFileListFromNodeResponse()
+        elif requestHead == 'cloneNodeRequest':
+            return server.createCloneNodeResponse()
 
 
 class ThreadedTCPServer(socketserver.ThreadingMixIn, socketserver.TCPServer):
@@ -68,35 +70,44 @@ class ThreadedTCPServer(socketserver.ThreadingMixIn, socketserver.TCPServer):
 
 
 class StorageNode(Server, baseClient):
-    def __init__(self, id, name, address, output):
+    def __init__(self, id, name, address, primaryServerAddress, output):
         self.statisticHelper = statisticHelper()
         Server.__init__(self, id, name, address, self.statisticHelper, ThreadedTCPServer, ThreadedTCPRequestHandler,
                         output)
         baseClient.__init__(self, id, name, address, self.statisticHelper, output)
-        self.nodeList = list()
+        self.cachedDirectoryServerAddress = primaryServerAddress
+        self.backupDirectoryServerAddress = None
         self.fileList = list()
-
-    def setNodeList(self, nodeList):
-        self.nodeList = nodeList
 
     def setFileList(self, fileList):
         self.fileList = fileList
 
-    def initFilePath(self):
-        if not self.getDirectoryPath().parent.exists():
-            os.mkdir(self.getDirectoryPath().parent)
-        if not self.getDirectoryPath().exists():
-            os.mkdir(self.getDirectoryPath())
+    def setup(self):
+        self.cleanFileDirectory()
+        self.initFilePath()
+        self.requestJoinNetwork()
+        self.getBackupServer()
+        self.requestCloneNode()
+
+
+    def switchBackupServer(self):
+        self.cachedDirectoryServerAddress = self.backupDirectoryServerAddress
+        self.backupDirectoryServerAddress = self.getBackupServer()
 
     def createAddFileResponse(self, fileName, content, forward):
         targetFilePath = self.getDirectoryPath().joinpath(fileName)
-        if not targetFilePath.exists():
+        if not targetFilePath.exists() and fileName not in self.fileList:
             with targetFilePath.open('wb') as file:
                 file.write(bytes(content, 'utf-8'))
             self.fileList.append(fileName)
         if forward:
-            self.sendMessage(DirectoryServer.getDirectoryServerAddress(),
-                             RequestAssembler.assembleNewFileRequest(fileName, content))
+            try:
+                self.sendMessage(self.cachedDirectoryServerAddress,
+                                 RequestAssembler.assembleNewFileRequest(fileName, content))
+            except ServerError:
+                self.switchBackupServer()
+                self.sendMessage(self.cachedDirectoryServerAddress,
+                                 RequestAssembler.assembleNewFileRequest(fileName, content))
         return ResponseAssembler.assembleAddFileResponse(True)
 
     def createReadFileResponse(self, fileName):
@@ -110,9 +121,15 @@ class StorageNode(Server, baseClient):
         return ResponseAssembler.assembleGetFileListFromNodeResponse(self.fileList)
 
     def requestJoinNetwork(self):
-        rawResponse = self.sendMessage(DirectoryServer.getDirectoryServerAddress(),
-                                       RequestAssembler.assembleJoinNetworkRequest(self.id, self.address[0],
-                                                                                   self.address[1]))
+        try:
+            rawResponse = self.sendMessage(self.cachedDirectoryServerAddress,
+                                           RequestAssembler.assembleJoinNetworkRequest(self.id, self.address[0],
+                                                                                       self.address[1]))
+        except ServerError:
+            self.switchBackupServer()
+            rawResponse = self.sendMessage(self.cachedDirectoryServerAddress,
+                                           RequestAssembler.assembleJoinNetworkRequest(self.id, self.address[0],
+                                                                                       self.address[1]))
         response = json.loads(rawResponse)
         if response['result'] != True:
             raise Exception('Join Network Error')
@@ -121,3 +138,38 @@ class StorageNode(Server, baseClient):
         with self.getDirectoryPath().joinpath(fileName).open('r') as file:
             content = file.read()
             return content
+
+    def requestGetBackupServer(self):
+        try:
+            return self.sendMessage(self.cachedDirectoryServerAddress,
+                                    RequestAssembler.assembleGetBackupServerRequest())
+        except ServerError:
+            self.switchBackupServer()
+            return self.sendMessage(self.cachedDirectoryServerAddress,
+                                    RequestAssembler.assembleGetBackupServerRequest())
+
+    def getBackupServer(self):
+        rawResponse = self.requestGetBackupServer()
+        response = json.loads(rawResponse)
+        self.backupDirectoryServerAddress = (response['serverIp'], response['serverPort'])
+
+    def requestCloneNode(self):
+        rawResponse = self.sendMessage(self.cachedDirectoryServerAddress, RequestAssembler.assembleCloneNodeRequest())
+        response = json.loads(rawResponse)
+        fileDict = response['fileDict']
+        cachedDirectoryServer = response['cachedDirectoryServer']
+        backupDirectoryServer = response['backupDirectoryServer']
+        self.fileList = list(fileDict.keys())
+        self.cachedDirectoryServerAddress = tuple(cachedDirectoryServer)
+        self.backupDirectoryServerAddress = tuple(backupDirectoryServer)
+        for fileName in self.fileList:
+            with self.getDirectoryPath().joinpath(fileName).open('wb') as file:
+                file.write(bytes(fileDict[fileName], 'utf-8'))
+
+    def createCloneNodeResponse(self):
+        fileDict = dict()
+        for fileName in self.fileList:
+            content = self.getFileContent(fileName)
+            fileDict[fileName] = content
+        return ResponseAssembler.assembleCloneNodeResponse(fileDict, self.cachedDirectoryServerAddress,
+                                                           self.backupDirectoryServerAddress)
